# docs/openapi.yaml
# used swagger rather than OpenAPI to generate the schema
swagger: "2.0"
basePath: /
paths:
  /api/location:
    parameters:
      - in: query
        description: Longitude of the location
        name: longitude
        type: string
      - in: query
        description: Latitude of the Location
        name: latitude
        type: string
      - in: query
        description: Time Location was Recorded
        name: creation_time
        type: string
      - in: query
        description: ID for Related Person
        name: person_id
        type: string
    post:
      responses:
        '200':
          description: Success
      operationId: post_location_create_resource
      tags:
        - UdaConnect
  /api/location/{location_id}:
    parameters:
      - type: string
        in: path
        description: Unique ID for a given Location
        name: location_id
        required: true
    get:
      responses:
        '200':
          description: Success
      operationId: get_location_resources
      tags:
        - UdaConnect
  /api/locations:
    get:
      responses:
        '200':
          description: Success
      operationId: get_locations_resource
      tags:
        - UdaConnect
  /api/persons:
    get:
      responses:
        '200':
          description: Success
      operationId: get_persons_resource
      tags:
        - UdaConnect
    post:
      responses:
        '200':
          description: Success
      operationId: post_persons_resource
      tags:
        - UdaConnect
  /api/persons/{person_id}:
    parameters:
      - in: path
        description: Unique ID for a given Person
        name: person_id
        required: true
        type: string
    get:
      responses:
        '200':
          description: Success
      operationId: get_person_resource
      tags:
        - UdaConnect
  /api/persons/{person_id}/connection:
    parameters:
      - in: query
        description: Proximity to a given user in meters
        name: distance
        type: string
      - in: query
        description: Upper bound of date range
        name: end_date
        type: string
      - in: query
        description: Lower bound of date range
        name: start_date
        type: string
      - name: person_id
        in: path
        required: true
        type: string
    get:
      responses:
        '200':
          description: Success
      operationId: get_connection_data_resource
      tags:
        - UdaConnect
info:
  title: UdaConnect API
  version: 0.1.0
produces:
  - application/json
consumes:
  - application/json
tags:
  - name: UdaConnect
    description: Connections via geolocation.
responses:
  ParseError:
    description: When a mask can't be parsed
  MaskError:
    description: When any error occurs on mask